requirements.txt
requests~=2.28.1
pytest~=7.1.2
pytest-ordering~=0.6
PyYAML~=6.0
jsonpath~=0.82